+++ create_db.py
@@ -1,0 +1,4 @@
+from app import app, db
+
+with app.app_context():
+    db.create_all()
